增

+++ src/main/java/com/xxz/dao/UserInfoMapper.java
@@ -3,13 +3,13 @@ package com.xxz.dao;
 import com.xxz.model.UserInfo;
 
 public interface UserInfoMapper {
-    int deleteByPrimaryKey(Integer userId);
+    int deleteByPrimaryKey(Long userId);
 
     int insert(UserInfo record);
 
     int insertSelective(UserInfo record);
 
-    UserInfo selectByPrimaryKey(Integer userId);
+    UserInfo selectByPrimaryKey(Long userId);
 
     int updateByPrimaryKeySelective(UserInfo record);
